chore: move ast definitions into submodules
