+++ src/interfaces/ast/index.ts
@@ -1,4 +1,2 @@
-export * from './map-ast';
-export * from './map-ast.utils';
-export * from './profile-ast';
-export * from './profile-ast.utils';
+export * from './map';
+export * from './profile';
